+++ src/main/java/utility/exceptions/ValidationException.java
@@ -1,0 +1,5 @@
+package utility.exceptions;
+
+// custom exception for validation error
+public class ValidationException extends Exception {
+}